set up file structure and fleshed out student class
